# SourceCode/front-end/logsmart/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "back-end", "description": "", "license": { "name": "" }, "version": "1.1.0" },
	"paths": {
...
		"/auth/company/branches": {
			"get": {
				"tags": ["Company Management"],
				"summary": "Lists all branches for the user's company.",
				"operationId": "list_branches",
				"responses": {
					"200": {
						"description": "Branches retrieved successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ListBranchesResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			},
			"post": {
				"tags": ["Company Management"],
				"summary": "Creates a new branch for the user's company.",
				"operationId": "create_branch",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/CreateBranchRequest" } }
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Branch created successfully",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/BranchDto" } }
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden - only company manager can create branches",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"BranchDto": {
				"type": "object",
				"required": ["id", "company_id", "name", "address", "created_at"],
				"properties": {
					"address": { "type": "string" },
					"company_id": { "type": "string" },
					"created_at": { "type": "string", "format": "date-time" },
					"id": { "type": "string" },
					"name": { "type": "string" }
				}
			},
...
			"CreateBranchRequest": {
				"type": "object",
				"required": ["name", "address"],
				"properties": {
					"address": { "type": "string", "example": "123 Regent St, London" },
					"name": { "type": "string", "example": "London Office" }
				}
			},
...
			"ErrorResponse": {
				"type": "object",
				"required": ["error"],
				"properties": { "error": { "type": "string" } }
			},
...
			"ListBranchesResponse": {
				"type": "object",
				"required": ["branches"],
				"properties": {
					"branches": { "type": "array", "items": { "$ref": "#/components/schemas/BranchDto" } }
				}
			},
...
		},
		"securitySchemes": { "bearer_auth": { "type": "http", "scheme": "bearer" } }
	},
...
}